Move SalesEntryScreen to a separate directory
